+++ src/main/java/io/serialized/client/aggregate/EventBatch.java
@@ -6,18 +6,18 @@ import java.util.Optional;
 
 public class EventBatch {
 
-  private Long expectedVersion;
+  private Integer expectedVersion;
   private List<Event<?>> events;
 
   public EventBatch() {
   }
 
-  public EventBatch(List<Event<?>> events, Long expectedVersion) {
+  public EventBatch(List<Event<?>> events, Integer expectedVersion) {
     this.events = events;
     this.expectedVersion = expectedVersion;
   }
 
-  public Long expectedVersion() {
+  public Integer expectedVersion() {
     return expectedVersion;
   }
